# Refactor, simplify and improve rendering performance 