--- a/lib/hanami/view.rb
+++ b/lib/hanami/view.rb
@@ -532,6 +532,8 @@ module Hanami
     #
     # @api public
     def initialize
+      ensure_config
+
       @exposures = self.class.exposures.bind(self)
     end
 
@@ -559,8 +561,6 @@ module Hanami
     # @return [Rendered] rendered view object
     # @api public
     def call(format: config.default_format, context: config.default_context, **input)
-      ensure_config
-
       env = self.class.render_env(format: format, context: context)
       template_env = self.class.template_env(format: format, context: context)