--- a/lib/hanami/view.rb
+++ b/lib/hanami/view.rb
@@ -529,13 +529,11 @@ module Hanami
 
     private
 
-    # @api private
     def ensure_config
       raise UndefinedConfigError, :paths unless Array(config.paths).any?
       raise UndefinedConfigError, :template unless config.template
     end
 
-    # @api private
     def locals(rendering, input)
       exposures.(context: rendering.context, **input) do |value, exposure|
         if exposure.decorate? && value
@@ -546,14 +544,12 @@ module Hanami
       end
     end
 
-    # @api private
     def layout_locals(locals)
       locals.each_with_object({}) do |(key, value), layout_locals|
         layout_locals[key] = value if exposures[key].for_layout?
       end
     end
 
-    # @api private
     def layout?
       !!config.layout # rubocop:disable Style/DoubleNegation
     end