--- a/lib/hanami/view/context.rb
+++ b/lib/hanami/view/context.rb
@@ -32,18 +32,6 @@ module Hanami
       end
 
       # @api private
-      SKIP_COPY_IVARS = %i[@_rendering].freeze
-      private_constant :SKIP_COPY_IVARS
-
-      # @api private
-      private def initialize_copy(source)
-        ivars = source.instance_variables - SKIP_COPY_IVARS
-        ivars.each do |ivar|
-          instance_variable_set(ivar, source.instance_variable_get(ivar).dup)
-        end
-      end
-
-      # @api private
       def dup_for_rendering(rendering)
         dup.tap do |obj|
           obj.instance_variable_set(:@_rendering, rendering)