--- a/lib/hanami/view/decorated_attributes.rb
+++ b/lib/hanami/view/decorated_attributes.rb
@@ -63,8 +63,8 @@ module Hanami
               define_method name do
                 attribute = super()
 
-                if _render_env && attribute
-                  _render_env.part(name, attribute, **options)
+                if _rendering && attribute
+                  _rendering.part(name, attribute, **options)
                 else
                   attribute
                 end