--- a/lib/hanami/view/errors.rb
+++ b/lib/hanami/view/errors.rb
@@ -45,5 +45,11 @@ module Hanami
         super(msg)
       end
     end
+
+    class RenderingMissingError < StandardError
+      def message
+        "a +rendering+ must be provided"
+      end
+    end
   end
 end