--- a/lib/hanami/view/exposure.rb
+++ b/lib/hanami/view/exposure.rb
@@ -56,15 +56,15 @@ module Hanami
       end
 
       def for_layout?
-        options.fetch(:layout) { false }
+        options.fetch(:layout, false)
       end
 
       def decorate?
-        options.fetch(:decorate) { true }
+        options.fetch(:decorate, true)
       end
 
       def private?
-        options.fetch(:private) { false }
+        options.fetch(:private, false)
       end
 
       def default_value