--- a/lib/hanami/view/part_builder.rb
+++ b/lib/hanami/view/part_builder.rb
@@ -1,6 +1,5 @@
 # frozen_string_literal: true
 
-require "dry/core/equalizer"
 require_relative "part"
 
 module Hanami
@@ -9,7 +8,6 @@ module Hanami
     #
     # @api private
     class PartBuilder
-      include Dry::Equalizer(:namespace)
 
       attr_reader :namespace
       attr_reader :rendering