… benchmarks/comparative/hanami/templates/users.html.erb
… lib/hanami/view/path.rb
… lib/hanami/view/render_environment.rb
… lib/hanami/view/rendering_missing.rb
… lib/hanami/view/scope.rb
… lib/hanami/view/scope_builder.rb
… lib/hanami/view/tilt.rb
… spec/fixtures/integration/scopes/anonymous_scope.html.slim
… spec/fixtures/integration/scopes/class_named_scope_with_implicit_render.html.slim
… spec/fixtures/integration/scopes/named_scope_with_defaults.html.slim
… spec/fixtures/integration/scopes/named_scope_with_explicit_render.html.slim
… spec/fixtures/integration/scopes/named_scope_with_implicit_render.html.slim
… spec/fixtures/integration/template_engines/erbse/render_and_yield.html.erb
… spec/fixtures/integration/template_engines/erubi/render_and_yield.html.erb
… spec/fixtures/integration/template_engines/hamlit/render_and_yield.html.haml
… spec/fixtures/integration/template_engines/slim/render_and_yield.html.slim
… spec/fixtures/integration/testing/_feature_box.html.slim
… spec/fixtures/templates/parts_with_args.html.slim
… spec/fixtures/templates/users.html.slim
… spec/fixtures/templates/users/_tbody.html.slim
… spec/fixtures/templates_override/users.html.slim
… spec/integration/part/decorated_attributes_spec.rb
… spec/integration/part_builder_spec.rb
… spec/integration/scope_spec.rb
… spec/integration/template_engines/erbse_spec.rb
… spec/integration/template_engines/hamlit_spec.rb
… spec/integration/testing/testing_parts_spec.rb
… spec/integration/view/errors_spec.rb
… spec/spec_helper.rb
… spec/unit/context_helpers/content_helpers_spec.rb