@timriley 2023-02-25
Thanks again for checking this over, @solnic and @jodosha! I've just made a few tweaks, many based on your feedback (see the last 5 commits), along with some adjustments in hanami/hanami to make sure views continue to work there with the changes I've made to `Context` here: https://github.com/hanami/hanami/pull/1292.

I'm merging this now!